refactor: Phase 4 - Reorganize into category-based structure
Organize plugins into logical subdirectories:
- lsp/ (LSP, formatting, completions)
- ui/ (theme, statusline, notifications)
- editor/ (treesitter, editing enhancements)
- navigation/ (telescope, harpoon, trouble)
- git/, ai/, debug/, tools/, terminal/

Reorganize core config:
- config/options.lua (vim settings)
- config/keymaps.lua (global keymaps)
- config/utils/ (shared utilities)

All plugins remain functional, just better organized.
