+++ nvim/lua/plugins/ai/ai.lua
