+++ nvim/lua/plugins/debug/debug.lua
