+++ nvim/lua/plugins/editor/treesitter.lua
