+++ nvim/lua/plugins/navigation/trouble.lua
