+++ nvim/lua/plugins/terminal/ghostty.lua
