+++ nvim/lua/plugins/terminal/kitty.lua
